1.新增FloatEditor浮点参数编辑状态机，分层架构解耦编码器/编辑/显示 2.Vset/Iset/Pset迁移为float类型，移除/100.0f换算 3.统一移除文件_k后缀，OLED_UI/PID迁移至App层 4.编码器改为增量累加器轮询，消除动态重新绑定

+++ Src/BSP/Inc/Digital_Power.h
@@ -1,6 +1,6 @@
 #include <ch32v00X.h>
 #include "PID.h"
-#include "Timer_k.h"
+#include "Timer.h"
 #include "Event_Bus.h"
 
 #define LM25118_ENABLE GPIO_SetBits(GPIOC, GPIO_Pin_5);
@@ -27,9 +27,9 @@ typedef struct {
     float OVP_Thredhold;
     float OCP_Thredhold;
 
-    s32 Vset;
-    s32 Iset;
-    s32 Pset;
+    float Vset;
+    float Iset;
+    float Pset;
 
 } Digital_Power_Dev;
